requirements.txt
flask==3.1.0
flask_classful==0.16.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
pytest==8.3.5
pytest_loop==1.0.13
pytest_mock==3.14.0
python_dotenv==1.1.0
pywebview==5.3.2
sqlalchemy==2.0.35